docker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
onnxruntime==1.19.2
numpy==2.1.3
Pillow==10.4.0
python-multipart==0.0.12
